# Printing performance numbers for volume lifecycle operations